+++ docs/index.dart
@@ -9,30 +9,7 @@ import 'dart:html';
 import '../lib/tex.dart';
 // ignore: avoid_relative_lib_imports
 import '../lib/src/tab.dart' as tab;
-
-final examples = [
-  'f(x)=x^2',
-  '\\frac 1 {x+1}',
-  '\\frac{z_1}{z_2}',
-  '\\sin x',
-  '\\sin(x)',
-  "\\sqrt{x+1}",
-  "\\overline{abc}",
-  '\\NN^2',
-  '\\lim_{x\\to\\infty}\\frac1x',
-  "\\begin{matrix}\n  1 & 2 \\\\\n  3 & 4 \\\\\n\\end{matrix}",
-  "\\begin{pmatrix}\n  1 & 2 \\\\\n  3 & 4 \\\\\n\\end{pmatrix}^T",
-  "\\begin{bmatrix}\n  1 & 2 \\\\\n  3 & 4 \\\\\n\\end{bmatrix}",
-  "\\begin{Bmatrix}\n  1 & 2 \\\\\n  3 & 4 \\\\\n\\end{Bmatrix}",
-  "\\begin{vmatrix}\n  1+\\alpha & 2 \\\\\n  3 & 4 \\\\\n\\end{vmatrix}",
-  "\\begin{vmatrix}[lc]\n  1+\\alpha & 2 \\\\\n  3 & 4 \\\\\n\\end{vmatrix}",
-  "sgn(x)=\\begin{cases}\n  -1 & x<0 \\\\\n  1 & x>0 \\\\\n  0 & \\text{otherwise}\n\\end{cases}",
-  "\\int_2^5 x \\, dx",
-  "\\sum_{i=1}^5 i^2",
-  "\\prod_{i=1}^5 i^2",
-  "\\{1, 2, \\dots, 5\\}",
-  "\\left( \\frac1x \\right)^2",
-];
+import 'examples.dart';
 
 final customMacros = ['\\CC', '\\NN', '\\QQ', '\\RR', '\\ZZ'];
 
@@ -181,8 +158,15 @@ void showGlyphTests() {
   for (var texSrc in tab.table.keys) {
     var rowData = tab.table[texSrc] as Map<String, Object>;
 
-    if (texSrc == '\\sqrt') texSrc = '\\sqrt{}';
-    if (texSrc == "'") texSrc = "{}'";
+    if (texSrc == '\\dot' ||
+        texSrc == '\\hat' ||
+        texSrc == '\\sqrt' ||
+        texSrc == '\\tilde') {
+      texSrc += '{}';
+    }
+    if (texSrc == "'") {
+      texSrc = "{}'";
+    }
 
     var tex = TeX();
     var displayStyle = true;
@@ -223,8 +207,12 @@ void showGlyphTests() {
     tr.append(cell3);
 
     var cell4 = document.createElement('td') as TableCellElement;
-    cell4.innerHtml = (rowData["d"] as int).toString();
+    cell4.innerHtml = (rowData["h"] as int).toString();
     tr.append(cell4);
+
+    var cell5 = document.createElement('td') as TableCellElement;
+    cell5.innerHtml = (rowData["d"] as int).toString();
+    tr.append(cell5);
 
     glyphTable.append(tr);
   }